Redux. Асинхронность
