+++ src/pages/SignIn/SignIn.jsx
@@ -5,19 +5,17 @@ import { useRef, useState } from 'react';
 import { UserContext } from '../../Authorization.jsx';
 import { useContext } from 'react';
 import { useNavigate } from 'react-router-dom';
+import { getToken,login } from '../../Api.jsx';
 
 export const Login = () => {
 
   const navigate = useNavigate()
-
   const [email, setEmail] = useState('')
-    const [password, setPassword] = useState('')
-    const [error, setError] = useState('')
-    const [userData, setUserData] = useContext(UserContext)
-
-    const signInButtonRef = useRef(null)
-
-    async function handleSignIn({ email, password }) {
+  const [password, setPassword] = useState('')
+  const [error, setError] = useState('')
+  const [, setUserData] = useContext(UserContext)
+  const signInButtonRef = useRef(null)
+  async function handleSignIn({ email, password }) {
       if (email === '') {
           setError('Укажите почту')
           return
@@ -26,44 +24,36 @@ export const Login = () => {
           setError('Укажите пароль')
           return
       }
-
-      try {
-          const response = await fetch(
-              'https://skypro-music-api.skyeng.tech/user/login/',
-              {
-                  method: 'POST',
-                  body: JSON.stringify({
-                      email: email,
-                      password: password,
-                  }),
-                  headers: {
-                      'content-type': 'application/json',
-                  },
-              },
-          )
-
-          if (response.status === 400) {
-              setError(
-                  'Произошла ошибка с данными. Неверные логин или пароль',
-              )
-              return
-          } else if (response.status === 401) {
-              setError('Пользователь с таким email или паролем не найден')
-              return
-          } else if (response.status === 500) {
-              setError('Сервер не отвечает, попробуй позже')
-              return
-          }
-
-          const data = await response.json()
-          setUserData(data)
-          setUserData(data.username)
-          localStorage.setItem('user', JSON.stringify(userData))
-            navigate('/')
-      } catch (error) {
-          console.log(error)
+      signInButtonRef.current.disabled = true
+      const response = await login({ email, password })
+      if (response.status === 400) {
+          setError('Произошла ошибка с данными. Неверные логин или пароль')
+          signInButtonRef.current.disabled = false
+          return
+      } else if (response.status === 401) {
+          setError('Пользователь с таким email или паролем не найден')
+          signInButtonRef.current.disabled = false
+          return
+      } else if (response.status === 500) {
+          setError('Сервер не отвечает, попробуй позже')
+          signInButtonRef.current.disabled = false
+          return
       }
+      const data = await response.json()
+      setUserData(data.username)
+      localStorage.setItem('user', JSON.stringify(data.username))
+      signInButtonRef.current.disabled = false
+      navigate('/')
+
+      const tokenResponse = await getToken({ email, password })
+      const tokens = await tokenResponse.json()
+      const accessToken = tokens.access
+      const refreshToken = tokens.refresh
+
+      localStorage.setItem('accessToken', accessToken)
+      localStorage.setItem('refreshToken', refreshToken)
   }
+    
     return(
     <>
       <GlobalStyle />
