+++ src/pages/myTrack/myTrack.styles.js
@@ -1,4 +1,4 @@
-import { styled } from 'styled-components'
+import { styled } from 'styled-components';
 
 export const Wrapper = styled.div`
     width: 100%;
@@ -6,7 +6,6 @@ export const Wrapper = styled.div`
     overflow: hidden;
     background-color: #383838;
 `
-
 export const Container = styled.div`
     max-width: 1920px;
     height: 100vh;
@@ -15,16 +14,166 @@ export const Container = styled.div`
     background-color: #181818;
     display: flex;
     justify-content: center;
+`
+
+export const Main = styled.main`
+    -webkit-box-flex: 1;
+    -ms-flex: 1 1 auto;
+    flex: 1 1 auto;
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -ms-flex-wrap: wrap;
+    flex-wrap: wrap;
+    -webkit-box-pack: justify;
+    -ms-flex-pack: justify;
+    justify-content: space-between;
+`
+
+export const MainCenterblock = styled.div`
+    width: auto;
+    -webkit-box-flex: 3;
+    -ms-flex-positive: 3;
+    flex-grow: 3;
+    padding: 20px 40px 20px 111px;
+`
+
+export const CenterblockSearch = styled.div`
+    width: 100%;
+    border-bottom: 1px solid #4e4e4e;
+    margin-bottom: 51px;
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-orient: horizontal;
+    -webkit-box-direction: normal;
+    -ms-flex-direction: row;
+    flex-direction: row;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
     align-items: center;
 `
 
-export const Text = styled.div`
+export const SearchSvg = styled.svg`
+    width: 17px;
+    height: 17px;
+    margin-right: 5px;
+    stroke: #ffffff;
+    fill: transparent;
+`
+
+export const SearchText = styled.input`
+    -webkit-box-flex: 100;
+    -ms-flex-positive: 100;
+    flex-grow: 100;
+    background-color: transparent;
+    border: none;
+    padding: 13px 10px 14px;
     font-style: normal;
-    font-weight: 700;
-    font-size: 56px;
-    line-height: 56px;
+    font-weight: 400;
+    font-size: 16px;
+    line-height: 24px;
+    color: #ffffff;
+    &::placeholder {
+        background-color: transparent;
+        color: #ffffff;
+        font-style: normal;
+        font-weight: 400;
+        font-size: 16px;
+        line-height: 24px;
+    }
+    &::-ms-input-placeholder {
+        background-color: transparent;
+        color: #ffffff;
+        font-style: normal;
+        font-weight: 400;
+        font-size: 16px;
+        line-height: 24px;
+    }
+    &:-ms-input-placeholder {
+        background-color: transparent;
+        color: #ffffff;
+        font-style: normal;
+        font-weight: 400;
+        font-size: 16px;
+        line-height: 24px;
+    }
+    &::-webkit-input-placeholder {
+        background-color: transparent;
+        color: #ffffff;
+        font-style: normal;
+        font-weight: 400;
+        font-size: 16px;
+        line-height: 24px;
+    }
+`
+
+export const CenterblockHeading = styled.h2`
+    font-style: normal;
+    font-weight: 400;
+    font-size: 64px;
+    line-height: 72px;
+    letter-spacing: -0.8px;
+    margin-bottom: 45px;
+`
+
+export const CenterblockContent = styled.div`
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-orient: vertical;
+    -webkit-box-direction: normal;
+    -ms-flex-direction: column;
+    flex-direction: column;
+`
+
+export const ContentTitle = styled.div`
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-orient: horizontal;
+    -webkit-box-direction: normal;
+    -ms-flex-direction: row;
+    flex-direction: row;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
+    align-items: center;
+    -webkit-box-pack: justify;
+    -ms-flex-pack: justify;
+    justify-content: space-between;
+    margin-bottom: 24px;
+`
+
+export const PlaylistTitle = styled.div`
+    font-style: normal;
+    font-weight: 400;
+    font-size: 14px;
+    line-height: 24px;
     letter-spacing: 2px;
-    color: #fff;
+    color: #696969;
     text-transform: uppercase;
-    text-align: center;
+`
+
+export const PlaylistTitleTrack = styled(PlaylistTitle)`
+    width: 447px;
+`
+
+export const PlaylistTitleAuthor = styled(PlaylistTitle)`
+    width: 321px;
+`
+
+export const PlaylistTitleAlbum = styled(PlaylistTitle)`
+    width: 245px;
+`
+
+export const PlaylistTitleImage = styled(PlaylistTitle)`
+    width: 60px;
+    text-align: end;
+`
+
+export const PlaylistTitleSvg = styled.svg`
+    width: 12px;
+    height: 12px;
+    fill: transparent;
+    stroke: #696969;
 `
