+++ src/pages/signUp/SignUp.jsx
@@ -5,18 +5,16 @@ import { useState, useRef } from 'react';
 import { UserContext } from '../../Authorization.jsx';
 import { useContext } from 'react';
 import { useNavigate } from 'react-router-dom';
+import { register, getToken } from '../../Api.jsx';
 
 export const SignUp = () => {
   const navigate = useNavigate()
-
-  const [email, setEmail] = useState('')
+    const [email, setEmail] = useState('')
     const [password, setPassword] = useState('')
     const [repeatPassword, setRepeatPassword] = useState('')
     const [error, setError] = useState('')
-    const [userData, setUserData] = useContext(UserContext)
-
+    const [, setUserData] = useContext(UserContext)
     const signUpButtonRef = useRef(null)
-
     async function handleSignUp({ email, password }) {
         if (email === '') {
             setError('Укажите почту')
@@ -30,40 +28,32 @@ export const SignUp = () => {
             setError('Пароли не совпадают')
             return
         }
-
-        try {
-            const response = await fetch(
-                'https://skypro-music-api.skyeng.tech/user/signup/',
-                {
-                    method: 'POST',
-                    body: JSON.stringify({
-                        email: email,
-                        password: password,
-                        username: email,
-                    }),
-                    headers: {
-                        'content-type': 'application/json',
-                    },
-                },
+        signUpButtonRef.current.disabled = true
+        const response = await register({ email, password })
+        if (response.status === 400) {
+            setError(
+                'Произошла ошибка с данными. Попробуйте изменить почту или пароль',
             )
-
-            if (response.status === 400) {
-                setError(
-                    'Вы вели неверные данные',
-                )
-                return
-            } else if (response.status === 500) {
-                setError('Сервер не отвечает, попробуй позже')
-                return
-            }
-
-            const data = await response.json()
-            setUserData(data.username)
-            localStorage.setItem('user', JSON.stringify(userData))
-            navigate('/')
-        } catch (error) {
-            console.log(error)
+            signUpButtonRef.current.disabled = false
+            return
+        } else if (response.status === 500) {
+            setError('Сервер не отвечает, попробуй позже')
+            signUpButtonRef.current.disabled = false
+            return
         }
+        const data = await response.json()
+        setUserData(data.username)
+        localStorage.setItem('user', JSON.stringify(data.username))
+        signUpButtonRef.current.disabled = false
+        navigate('/')
+
+        const tokenResponse = await getToken({ email, password })
+        const tokens = await tokenResponse.json()
+        const accessToken = tokens.access
+        const refreshToken = tokens.refresh
+
+        localStorage.setItem('accessToken', accessToken)
+        localStorage.setItem('refreshToken', refreshToken)
     }
     
     return(
