+++ src/store/index.jsx
@@ -1,5 +1,5 @@
-import { configureStore } from '@reduxjs/toolkit'
-import { trackReducer } from './slices'
+import { configureStore } from '@reduxjs/toolkit';
+import { trackReducer } from './slices';
 
 export default configureStore({
     reducer: {
