+++ src/store/slices.jsx
@@ -5,6 +5,7 @@ const trackSlice = createSlice({
     initialState: {
         activeTrack: {},
         tracks: [],
+        currentPlaylist: [],
         shuffledTracks: [],
         isShuffled: false,
     },
@@ -17,6 +18,10 @@ const trackSlice = createSlice({
             )
         },
 
+        setCurrentPlaylist(state) {
+            state.currentPlaylist = state.tracks
+        },
+
         setActiveTrack(state, action) {
             state.activeTrack = action.payload.track
         },
@@ -24,6 +29,7 @@ const trackSlice = createSlice({
         setIsShuffled(state) {
             state.isShuffled = !state.isShuffled
             if (state.isShuffled) {
+                state.shuffledTracks = state.currentPlaylist
                 state.shuffledTracks.sort(() => Math.random() - 0.5)
             }
         },
@@ -31,7 +37,7 @@ const trackSlice = createSlice({
         playNextTrack(state) {
             const playlist = state.isShuffled
                 ? state.shuffledTracks
-                : state.tracks
+                : state.currentPlaylist
             const indexCurrentTrack = playlist.findIndex((track) => {
                 return track.id === state.activeTrack.id
             })
@@ -44,7 +50,7 @@ const trackSlice = createSlice({
         playPrevTrack(state) {
             const playlist = state.isShuffled
                 ? state.shuffledTracks
-                : state.tracks
+                : state.currentPlaylist
 
             const indexCurrentTrack = playlist.findIndex((track) => {
                 return track.id === state.activeTrack.id
@@ -63,6 +69,7 @@ export const {
     playNextTrack,
     playPrevTrack,
     setTracks,
+    setCurrentPlaylist,
 } = trackSlice.actions
 
 export const trackReducer = trackSlice.reducer
